requirements.txt
python-dotenv
langgraph
langchain
langchain-openai
pydantic
typing_extensions
python-docx
streamlit
